+++ semcore/base-trigger/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.0.0] - unreleased
+
+### BREAK
+
+- Restyling component.
+- Remove value "xl" and "s" for "size".
+
 ## [2.6.3] - 2022-03-21
 
 ### Changed